simplify server configuration structure
