+++ scripts/post-build-prod.js
@@ -12,7 +12,7 @@ var rootPath = findRoot(process.cwd());
 try {
 
   var distLocalMachineFolder =  path.resolve(rootPath, 'dist/__local_machine_only__');
-    var serverConfigPath = path.resolve(rootPath,'dist/kmc-config.json');
+    var serverConfigPath = path.resolve(rootPath,'dist/server-config.json');
 
   if (fs.existsSync(distLocalMachineFolder)) {
     console.log('deleting folder "dist/__local_machine_only__" (should be used for development purposes only)');
@@ -20,13 +20,13 @@ try {
   }
 
     if (fs.existsSync(serverConfigPath)) {
-        console.log('deleting file "dist/kmc-config.json" (should be used for development purposes only)');
+        console.log('deleting file "dist/server-config.json" (should be used for development purposes only)');
 
         rimraf.sync(serverConfigPath);
     }
 
 }catch(e)
 {
-	console.error('ERROR! failed to delete "dist/__local_machine_only__" and "dist/kmc-config.json" (should be used for development purposes only)');
+	console.error('ERROR! failed to delete "dist/__local_machine_only__" and "dist/server-config.json" (should be used for development purposes only)');
 	throw e
 }
