+++ src/.gitignore
@@ -1,1 +1,1 @@
-kmc-config.json
+server-config.json
