+++ src/configuration/server-config.ts
@@ -1,181 +1,23 @@
 import 'rxjs/add/operator/takeUntil';
 import 'rxjs/add/operator/delay';
 import {globalConfig} from './global-config';
+import {Observable} from 'rxjs/Observable';
+import {environment} from 'environments/environment';
+import * as Ajv from 'ajv';
+import { ServerConfigSchema } from './server-config-schema';
 
 /*************************************
  * Developer Notice:
  * When you add/edit/remove server configuration you should sync the following places:
- * - this file > 'ServerConfigSchema' constant
+ * - server-config-schema.ts > 'ServerConfigSchema' constant
  * - this file > 'ServerConfig' interface
- * - file 'server-config.template.json' (used by CI server)
+ * - server-config.example.json (used during development)
+ * - server-config.template.json (used by CI server)
  *
  * If you are modifing an External application (a.k.a standalone application):
  * - file '__local_machine_only__/README.md' > section 'Test external applications integration'
  * - for new external apps you should also update the zip file 'samples-for-tests-only.zip'
  *************************************/
-
-export const ServerConfigSchema = {
-    properties: {
-        kalturaServer: {
-            properties: {
-                uri: {type: 'string'},
-                previewUIConf: {type: 'number'},
-                freeTrialExpiration: {
-                    properties: {
-                        enabled: {type: 'boolean'},
-                        trialPeriodInDays: {type: 'number'}
-                    },
-                    required: ['enabled', 'trialPeriodInDays'],
-                    additionalProperties: false
-                }
-
-            },
-            required: ['uri', 'previewUIConf', 'freeTrialExpiration'],
-            additionalProperties: false
-        },
-        cdnServers: {
-            properties: {
-                serverUri: {type: 'string'},
-                securedServerUri: {type: 'string'}
-            },
-            required: ['serverUri', 'securedServerUri'],
-            additionalProperties: false
-        },
-        externalApps: {
-            properties: {
-                studio: {
-                    properties: {
-                        enabled: {type: 'boolean'},
-                        uri: {type: 'string'},
-                        html5_version: {type: 'string'},
-                        html5lib: {type: 'string'},
-                        showStudioV3: {type: 'boolean'}
-                    },
-                    required: ['enabled', 'uri', 'html5_version', 'html5lib'],
-                    additionalProperties: false
-                },
-                studioV3: {
-                    properties: {
-                        enabled: {type: 'boolean'},
-                        uri: {type: 'string'},
-                        html5_version: {type: 'string'},
-                        html5lib: {type: 'string'},
-                        showHTMLStudio: {type: 'boolean'}
-                    },
-                    required: ['enabled', 'uri', 'html5_version', 'html5lib'],
-                    additionalProperties: false
-                },
-                usageDashboard: {
-                    properties: {
-                        enabled: {type: 'boolean'},
-                        uri: {type: 'string'},
-                        uiConfId: {type: 'number'},
-                        map_urls: { type: 'array', items: { type: 'string' } },
-                        map_zoom_levels: {type: 'string'}
-                    },
-                    required: ['enabled', 'uri', 'uiConfId', 'map_urls', 'map_zoom_levels'],
-                    additionalProperties: false
-                },
-                liveDashboard: {
-                    properties: {
-                        enabled: {type: 'boolean'},
-                        uri: {type: 'string'}
-                    },
-                    required: ['enabled', 'uri'],
-                    additionalProperties: false
-                },
-                liveAnalytics: {
-                    properties: {
-                        enabled: {type: 'boolean'},
-                        uri: {type: 'string'},
-                      uiConfId: {type: 'number'}
-                    },
-                    required: ['enabled', 'uri'],
-                    additionalProperties: false
-                },
-                clipAndTrim: {
-                  properties: {
-                    enabled: {type: 'boolean'},
-                    uri: {type: 'string'}
-                  },
-                  required: ['enabled', 'uri'],
-                  additionalProperties: false
-                },
-                advertisements: {
-                  properties: {
-                    enabled: {type: 'boolean'},
-                    uri: {type: 'string'}
-                  },
-                  required: ['enabled', 'uri'],
-                  additionalProperties: false
-                },
-                kava: {
-                    properties: {
-                        enabled: {type: 'boolean'},
-                        uri: {type: 'string'}
-                    },
-                    required: ['enabled', 'uri'],
-                    additionalProperties: false
-                }
-            },
-            required: ['studio', 'usageDashboard', 'liveDashboard', 'kava'],
-            additionalProperties: false
-        },
-        externalLinks: {
-            properties: {
-                previewAndEmbed: {
-                    properties: {
-                        embedTypes: {type: 'string'},
-                        deliveryProtocols: {type: 'string'}
-                    },
-                    required: ['embedTypes', 'deliveryProtocols'],
-                    additionalProperties: false
-                },
-                kaltura: {
-                    properties: {
-                        kmcOverview: {type: 'string'},
-                        mediaManagement: {type: 'string'},
-                        userManual: {type: 'string'},
-                        support: {type: 'string'},
-                        signUp: {type: 'string'},
-                        contactUs: {type: 'string'},
-                        upgradeAccount: {type: 'string'},
-                        contactSalesforce: {type: 'string'}
-                    },
-                    required: ['userManual', 'support', 'signUp', 'contactUs', 'upgradeAccount', 'contactSalesforce'],
-                    additionalProperties: false
-                },
-                entitlements: {
-                    properties: {
-                        manage: {type: 'string'}
-                    },
-                    required: ['manage'],
-                    additionalProperties: false
-                },
-                uploads: {
-                    properties: {
-                        highSpeedUpload: {type: 'string'},
-                        needHighSpeedUpload: {type: 'string'},
-                        bulkUploadSamples: {type: 'string'}
-                    },
-                    required: ['highSpeedUpload', 'needHighSpeedUpload', 'bulkUploadSamples'],
-                    additionalProperties: false
-                },
-                live: {
-                    properties: {
-                        akamaiEdgeServerIpURL: {type: 'string'}
-                    },
-                    required: ['akamaiEdgeServerIpURL'],
-                    additionalProperties: false
-                }
-            },
-            required: ['previewAndEmbed', 'kaltura', 'uploads', 'live'],
-            additionalProperties: false
-        }
-    },
-    required: ['kalturaServer', 'cdnServers', 'externalApps', 'externalLinks'],
-    additionalProperties: false
-};
 
 export interface ServerConfig {
     kalturaServer: {
@@ -239,7 +81,7 @@ export interface ServerConfig {
             embedTypes: string,
             deliveryProtocols: string
         },
-        entitlements:{
+        entitlements: {
             manage: string
         },
         kaltura: {
@@ -273,4 +115,90 @@ export function getKalturaServerUri(suffix: string = ''): string {
     } else {
         throw new Error(`cannot provide kaltura server uri. server configuration wasn't loaded already`);
     }
+}
+
+
+function validateSeverConfig(data: ServerConfig): { isValid: boolean, error?: string } {
+    const ajv = new Ajv({allErrors: true, verbose: true});
+    const validate = ajv.compile(ServerConfigSchema);
+    const isValid = !!validate(data);
+    let error = null;
+
+    if (!isValid) {
+        error = ajv.errorsText(validate.errors);
+    }
+
+    return { isValid, error };
+}
+
+
+function getConfiguration(): Observable<ServerConfig> {
+    if (window && (<any>window).kmcConfig) {
+        return Observable.of((<any>window).kmcConfig);
+    }
+
+    return Observable.create(observer =>
+    {
+        let completed = false;
+        const xhr = new XMLHttpRequest();
+
+        xhr.onreadystatechange = function () {
+            if (xhr.readyState === 4) {
+                let resp;
+
+                completed = true;
+
+                try {
+                    if (xhr.status === 200) {
+                        resp = JSON.parse(xhr.response);
+                    } else {
+                        if (globalConfig.client.production) {
+                            resp = new Error('failed to load configuration file from server with error ' + xhr.statusText);
+                        }else {
+                            resp = new Error('failed to load configuration file from server with error ' + xhr.statusText + ' (did you remember to create a configuration file from the provided template in the app folder?)');
+                        }
+
+                    }
+                } catch (e) {
+                    resp = e;
+                }
+
+                if (resp instanceof Error) {
+                    observer.error(resp);
+                } else {
+                    observer.next(resp);
+                    observer.complete();
+                }
+            }
+        };
+
+        xhr.open('Get', `${environment.configurationUri}?v=${globalConfig.client.appVersion}`);
+
+        xhr.send();
+
+        return () =>
+        {
+            if (!completed) {
+                console.warn('request to get application configuration was aborted');
+                xhr.abort();
+            }
+        }
+    });
+}
+
+export function initializeConfiguration(): Observable<void> {
+
+    return getConfiguration()
+        .takeUntil(Observable.of(true).delay(environment.configurationTimeout))
+        .do(response => {
+            const validationResult = validateSeverConfig(response);
+            if (validationResult.isValid) {
+                Object.assign(serverConfig, response);
+            } else {
+                throw Error(validationResult.error || 'Invalid server configuration');
+            }
+        })
+        .map(() => {
+            return undefined;
+        });
 }
