+++ src/environments/environment.ts
@@ -11,7 +11,7 @@
 
 export const environment = {
   production: false,
-    configurationUri: 'kmc-config.json',
+    configurationUri: 'server-config.json',
     configurationTimeout: 10000,
     client: {
       useSecuredProtocol: false
