+++ src/shared/kmc-shared/kmc-views/details-views/content-entry-view.service.ts
@@ -34,6 +34,7 @@ export interface ContentEntryViewArgs {
     entry: KalturaMediaEntry;
     section?: ContentEntryViewSections;
     activatedRoute?: ActivatedRoute;
+    reloadEntriesListOnNavigateOut?: boolean;
 }
 
 
@@ -199,7 +200,7 @@ export class ContentEntryViewService extends KmcDetailsViewBaseService<ContentEn
     protected _open(args: ContentEntryViewArgs): Observable<boolean> {
         const sectionToken = this._getSectionRouteToken(args.section);
         this._logger.info('handle open entry view request by the user', { entryId: args.entry.id, sectionToken });
-        return Observable.fromPromise(this._router.navigateByUrl(`/content/entries/entry/${args.entry.id}/${sectionToken}`));
+        return Observable.fromPromise(this._router.navigateByUrl(`/content/entries/entry/${args.entry.id}/${sectionToken}`, {queryParams: {reloadEntriesListOnNavigateOut: args.reloadEntriesListOnNavigateOut}}));
     }
 
     public openById(entryId: string): Observable<boolean> {
